requirements.txt
streamlit
boto3
elasticsearch
google-generativeai
python-dotenv
